[Azoth Sarin] Store the file size as quint64.
Tox implementation does it this way, so let's mirror its behaviour.

+++ src/plugins/azoth/plugins/sarin/filetransfermanager.cpp
@@ -161,7 +161,7 @@ namespace Sarin
 			pkey,
 			friendNum,
 			filenum,
-			static_cast<qint64> (size),
+			size,
 			name,
 			toxThread
 		};
